% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Bed_Split}
\alias{Bed_Split}
\title{Bed_Split}
\usage{
Bed_Split(inputBedFile, readLength)
}
\arguments{
\item{inputBedFile}{Input BED file that is not strand split. For PIPETS, the first column must be the chromosome name, the second column must be the start coordinate, the third column must be the stop coordinate, and the 6th column must be the strand. Columns 4 and 5 must be present but their information will not be used.}

\item{readLength}{The user must input the expected length of each “proper” read from the 3’-seq protocol. PIPETS gives a +/- 2 range for detecting reads to be used in the input BED files based on distributions of read coverage in parameter testing.}
}
\value{
Returns a list containing the Plus Strand Reads, the Minus Strand Reads, and the user defined name for the files. Also writes out the strand split bed files to the project directory.
}
\description{
First step of PIPETS. Takes input Bed files and splits them by strand while also assigning read coverage to each genomic position
}
