% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CompStrand_InitialCondense}
\alias{CompStrand_InitialCondense}
\title{CompStrand_InitialCondense}
\usage{
CompStrand_InitialCondense(CompInitialPoisson, adjacentPeakDistance = 2)
}
\arguments{
\item{CompInitialPoisson}{Uses the output of CompStrand_InitialPoisson as the input}

\item{adjacentPeakDistance}{adjacentPeakDistance During the peak condensing step, this parameter is used to define “adjacent” for significant genomic positions. This is used to identify initial peak structures in the data. By default this value is set to 2 to ensure that single instances of loss of signal are not sufficient to prevent otherwise contiguous peak signatures from being combined.}
}
\value{
Returns a dataframe that contains the list of termination "peaks" for the complement strand
}
\description{
Takes the significant top strand positions from CompStrand_InitialPoisson and condenses all proximal positions into termination "peaks"
}
