% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CompStrand_InitialPoisson}
\alias{CompStrand_InitialPoisson}
\title{CompStrand_InitialPoisson}
\usage{
CompStrand_InitialPoisson(
  PlusStrandReads,
  slidingWindowSize = 25,
  slidingWindowMovementDistance = 25,
  threshAdjust = 0.75,
  pValue = 0.005,
  topEndPercentage = 0.01
)
}
\arguments{
\item{PlusStrandReads}{Plus Strand Read DataFrame from the Bed_Split method. The Plus strand reads inform the Complement Strand termination signal}

\item{slidingWindowSize}{This parameter establishes the distance up and down stream of each position that a sliding window will be created around. The default value is 25, and this will result in a sliding window of total size 51 (25 upstream + position (1) + 25 downstream).}

\item{slidingWindowMovementDistance}{This parameter sets the distance that the sliding window will be moved. By default, it is set to move by half of the sliding window size in order to ensure that almost every position in the data is tested twice.}

\item{threshAdjust}{This parameter is used to establish a global cutoff threshold informed by the data. PIPETS sorts the genomic positions of each strand from highest to lowest, and starts with the highest read coverage position and subtracts that value from the total read coverage for that strand. By default, this continues until 75\% of the total read coverage has been accounted for. Increasing the percentage (e.x. 0.9) will lower the strictness of the cutoff, thus increasing the total number of significant results.}

\item{pValue}{Choose the minimum pValue that the Poisson distribution test must pass in order to be considered significant}

\item{topEndPercentage}{This parameter is used along with threshAdjust to trim off the influence exerted by high read coverage outliers. By default, it removes the top 0.01 percent of the highest read coverage positions from the calculation of the global threshold (e.x. if there are 200 positions that make up 75\% of the total reads, then this parameter will take the top 2 read coverage positions and remove them from the calculation of the global threshold). This parameter can be tuned to account for datasets with outliers that would otherwise severely skew the global threshold.}
}
\value{
Returns a dataframe with all genomic positions that were identified as having significant read coverage.
}
\description{
Poisson Significant Peak Identification Test for the Complement Strand Data
}
