% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CompStrand_SecondaryCondense}
\alias{CompStrand_SecondaryCondense}
\title{CompStrand_SecondaryCondense}
\usage{
CompStrand_SecondaryCondense(
  CompInitialCondense,
  peakCondensingDistance = 20,
  OutputFileName
)
}
\arguments{
\item{CompInitialCondense}{Uses the output of CompStrand_InitialCondense as the input}

\item{peakCondensingDistance}{peakCondensingDistance Following the initial peak condensing step, this parameter is used to identify peak structures in the data that are close enough to be considered part of the same termination signal. In testing, we have not identified cases in which two distinct termination signals so proximal that the default parameters incorrectly combine the signals together.}

\item{OutputFileName}{A string that will be used to identify printed results files. When run with PIPETS_FullRun, this string will be input by the user in the beginning}
}
\value{
The method writes a csv file in the project directory that contains the results for the Complement Strand termination peaks
}
\description{
Takes the list of termination "peaks" from CompStrand_InitialCondense and condenses peaks that are proximal to each other
}
