% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{consecutiveCheck}
\alias{consecutiveCheck}
\title{consecutiveCheck}
\usage{
consecutiveCheck(OF, OMF, aPD)
}
\arguments{
\item{OF}{Dataframe with significant positions (after initial Poisson test)}

\item{OMF}{Dataframe that will be the output, specified in function this is nested in}

\item{aPD}{Adjacent Peak distance speicified by user in full run}
}
\value{
Returns list of merged termination paeks
}
\description{
Identifies significant positions that are consecutive
}
