% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{consecutivePeakCheck}
\alias{consecutivePeakCheck}
\title{consecutivePeakCheck}
\usage{
consecutivePeakCheck(OMF, SWR, pCD)
}
\arguments{
\item{OMF}{OutputMergeFrame from function this is nested in}

\item{SWR}{Output of file that is created in nested function}

\item{pCD}{peak condensing distance established in full run}
}
\value{
Returns significant peaks to be fixed and output after
}
\description{
Combines Proximal Peaks
}
