% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{inputCheck}
\alias{inputCheck}
\title{inputCheck}
\usage{
inputCheck(
  inputBedFile,
  readLength,
  slidingWindowSize,
  slidingWindowMovementDistance,
  threshAdjust,
  pValue,
  topEndPercentage,
  adjacentPeakDistance,
  peakCondensingDistance
)
}
\arguments{
\item{inputBedFile}{See PIPETS_Run for full explanation}

\item{readLength}{See PIPETS_Run for full explanation}

\item{slidingWindowSize}{See PIPETS_Run for full explanation}

\item{slidingWindowMovementDistance}{See PIPETS_Run for full explanation}

\item{threshAdjust}{See PIPETS_Run for full explanation}

\item{pValue}{See PIPETS_Run for full explanation}

\item{topEndPercentage}{See PIPETS_Run for full explanation}

\item{adjacentPeakDistance}{See PIPETS_Run for full explanation}

\item{peakCondensingDistance}{See PIPETS_Run for full explanation}
}
\value{
Returns kicker variable that will stop PIPETS if error is detected
}
\description{
Checks user inputs to ensure PIPETS can run with given parameters
}
