% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{thresCalc}
\alias{thresCalc}
\title{thresCalc}
\usage{
thresCalc(rf, threshAdjust, topEndPercentage)
}
\arguments{
\item{rf}{Dataframe containing strand specific reads}

\item{threshAdjust}{This parameter is used to establish a global cutoff threshold informed by the data. PIPETS sorts the genomic positions of each strand from highest to lowest, and starts with the highest read coverage position and subtracts that value from the total read coverage for that strand. By default, this continues until 75\% of the total read coverage has been accounted for. Increasing the percentage (e.x. 0.9) will lower the strictness of the cutoff, thus increasing the total number of significant results.}

\item{topEndPercentage}{This parameter is used along with threshAdjust to trim off the influence exerted by high read coverage outliers. By default, it removes the top 0.01 percent of the highest read coverage positions from the calculation of the global threshold (e.x. if there are 200 positions that make up 75\% of the total reads, then this parameter will take the top 2 read coverage positions and remove them from the calculation of the global threshold). This parameter can be tuned to account for datasets with outliers that would otherwise severely skew the global threshold.}
}
\value{
Outputs threshold used for cutoff
}
\description{
Used to calculate a cutoff threshold for the data
}
