% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PIPETS_FullRun}
\alias{PIPETS_FullRun}
\title{PIPETS_FullRun}
\usage{
PIPETS_FullRun(
  inputBedFile,
  readLength,
  slidingWindowSize = 25,
  slidingWindowMovementDistance = 25,
  threshAdjust = 0.75,
  pValue = 0.005,
  topEndPercentage = 0.01,
  adjacentPeakDistance = 2,
  peakCondensingDistance = 20
)
}
\arguments{
\item{inputBedFile}{Input BED file that is not strand split. For PIPETS, the first column must be the chromosome name, the second column must be the start coordinate, the third column must be the stop coordinate, and the 6th column must be the strand. Columns 4 and 5 must be present but their information will not be used.}

\item{readLength}{The user must input the expected length of each “proper” read from the 3’-seq protocol. PIPETS gives a +/- 2 range for detecting reads to be used in the input BED files based on distributions of read coverage in parameter testing.}

\item{slidingWindowSize}{This parameter establishes the distance up and down stream of each position that a sliding window will be created around. The default value is 25, and this will result in a sliding window of total size 51 (25 upstream + position (1) + 25 downstream).}

\item{slidingWindowMovementDistance}{This parameter sets the distance that the sliding window will be moved. By default, it is set to move by half of the sliding window size in order to ensure that almost every position in the data is tested twice.}

\item{threshAdjust}{This parameter is used to establish a global cutoff threshold informed by the data. PIPETS sorts the genomic positions of each strand from highest to lowest, and starts with the highest read coverage position and subtracts that value from the total read coverage for that strand. By default, this continues until 75\% of the total read coverage has been accounted for. Increasing the percentage (e.x. 0.9) will lower the strictness of the cutoff, thus increasing the total number of significant results.}

\item{pValue}{Choose the minimum pValue that the Poisson distribution test must pass in order to be considered significant}

\item{topEndPercentage}{This parameter is used along with threshAdjust to trim off the influence exerted by high read coverage outliers. By default, it removes the top 0.01 percent of the highest read coverage positions from the calculation of the global threshold (e.x. if there are 200 positions that make up 75\% of the total reads, then this parameter will take the top 2 read coverage positions and remove them from the calculation of the global threshold). This parameter can be tuned to account for datasets with outliers that would otherwise severely skew the global threshold.}

\item{adjacentPeakDistance}{During the peak condensing step, this parameter is used to define “adjacent” for significant genomic positions. This is used to identify initial peak structures in the data. By default this value is set to 2 to ensure that single instances of loss of signal are not sufficient to prevent otherwise contiguous peak signatures from being combined.}

\item{peakCondensingDistance}{Following the initial peak condensing step, this parameter is used to identify peak structures in the data that are close enough to be considered part of the same termination signal. In testing, we have not identified cases in which two distinct termination signals so proximal that the default parameters incorrectly combine the signals together.}
}
\value{
PIPETS outputs strand specific results files as well as strand specific bed files to the directory that the R project is in.
}
\description{
Poisson Identification of PEaks from Term-Seq data. This is the full run method that begins with input Bed file and returns the strand split results
}
